# ENH: faster DXT interpolation