* apply the faster interpolation algorithm
described at:
https://github.com/darshan-hpc/darshan/issues/798#issuecomment-1226383841

* as I note in the source, this could likely be even
faster if we used Cython or numba, but I decided against
the complexity of a new build or runtime dependency just yet..

* this also has the advantage of removing a version
switch for Python (which actually should have been
a version switch for `pandas`)

* the HTML report generation time for `e3sm_io_heatmap_and_dxt.darshan` drops from `47` to `33` seconds on my machine, while the `snyder_acme..` log timing remains unchaged (dominated by "data ingestion" there); also, our internal benchmark suite shows this branch as decisively faster than `main`, so I don't really see a drawback:

`asv continuous -e -b ".*heatmap.*" main treddy_faster_dxt_interpolate`

```
       before           after         ratio
     [3fb01f49]       [9d11b152]
     <main>           <treddy_faster_dxt_interpolate>
-     13.0±0.09ms      9.88±0.09ms     0.76  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(50, 10, 0.01)
-      13.1±0.2ms       9.83±0.1ms     0.75  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(50, 10, 0.001)
-      13.2±0.1ms      9.80±0.04ms     0.74  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(1000, 10, 0.001)
-      14.5±0.1ms      10.0±0.09ms     0.69  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(10000, 10, 0.001)
-      14.6±0.1ms      9.87±0.03ms     0.68  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(1000, 10, 0.01)
-      31.1±0.2ms      16.6±0.08ms     0.53  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(50, 50, 0.001)
-      31.6±0.4ms       16.9±0.2ms     0.53  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(1000, 50, 0.001)
-      31.4±0.3ms      16.7±0.04ms     0.53  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(50, 50, 0.01)
-      33.4±0.3ms       17.3±0.2ms     0.52  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(10000, 50, 0.001)
-      33.2±0.4ms       16.9±0.1ms     0.51  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(1000, 50, 0.01)
-      19.7±0.1ms      9.93±0.04ms     0.50  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(50, 10, 1.0)
-      38.7±0.4ms       17.2±0.2ms     0.45  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(50, 50, 1.0)
-         124±2ms       53.2±0.3ms     0.43  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(10000, 250, 0.001)
-         120±2ms       50.4±0.3ms     0.42  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(1000, 250, 0.001)
-       120±0.7ms       50.5±0.4ms     0.42  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(50, 250, 0.01)
-         120±2ms       50.3±0.3ms     0.42  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(50, 250, 0.001)
-         123±1ms       51.4±0.3ms     0.42  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(1000, 250, 0.01)
-         141±3ms       57.5±0.6ms     0.41  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(10000, 250, 0.01)
-         130±2ms         52.7±1ms     0.40  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(50, 250, 1.0)
-      46.2±0.3ms      18.4±0.09ms     0.40  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(10000, 50, 0.01)
-      26.6±0.2ms      10.3±0.04ms     0.39  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(10000, 10, 0.01)
-       300±0.9ms        114±0.3ms     0.38  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(1000, 250, 1.0)
-      1.87±0.01s          660±1ms     0.35  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(10000, 250, 1.0)
-         167±3ms       28.8±0.3ms     0.17  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(1000, 50, 1.0)
-      1.37±0.01s        151±0.6ms     0.11  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(10000, 50, 1.0)
-       141±0.8ms       12.8±0.1ms     0.09  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(1000, 10, 1.0)
-      1.27±0.01s       41.2±0.6ms     0.03  dxt_heatmap.GetHeatMapDf.time_get_heatmap_df(10000, 10, 1.0)

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.

```