@tylerjereddy 2022-09-07
Ok, it has an approval and it fixes the CI problem after close/reopen by removing the transitive dependency on SciPy & making algo faster at the same time.